Refactor CSS height and line-height properties
Adjust the height and line-height properties for the .wrapped .dynamic-txts class in styling.css. The height is reduced from 100px to 75px, and the line-height is set to 85px. This change improves the layout of the dynamic text elements.

Add new SVG assets

Introduce two new SVG assets, github-svgrepo-com.svg and linkedin-svgrepo-com.svg, to the assets directory. These assets were uploaded to SVG Repo and will be used in the application.

Update linkedin-svgrepo-com.svg

Replace the existing linkedin-svgrepo-com.svg asset with an updated version. The new asset has a larger viewbox and improved styling.

Add linkedin-svgrepo-com (1).svg

Include an additional linkedin-svgrepo-com (1).svg asset to the assets directory. This asset is an alternative version of the LinkedIn logo.